Removed quotes from around the 0 in press ""0""
Some CSV parsers don't like the double quoting and fail on these
lines, some silently fail.

You could blame the parsers but CSV is not that well defined as a
format. It seems sensible to just remove the quotes rather than
have everyone try to code around it (and spend time finding
out what the problem is).